01-interactive-analysis/pyproject.toml
[project]
name = "presentation-demo-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fredapi>=0.5.2",
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
]
